requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-bootstrap4==3.0.1
django-crispy==1.8.1
django-crispy-forms==1.11.0
django-dotenv==1.4.2
django-etc==1.3.0
django-hitcount==1.3.3
django-resized==0.3.11
django-taggit==1.5.0
django-tinymce==3.3.0
gunicorn==20.1.0
importlib-metadata==3.10.1
Pillow==8.3.0
psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
zipp==3.5.0
